pyproject.toml
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iap-local-emulator"
version = "0.1.0"
description = "A complete local development environment for Google Play In-App Purchase and subscription lifecycle testing"
readme = "readme.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Google IAP Emulator Contributors"}
]
keywords = [
    "google-play",
    "iap",
    "in-app-purchase",
    "subscriptions",
    "emulator",
    "testing",
    "pubsub",
    "android"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Testing :: Mocking",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "google-cloud-pubsub>=2.19.0",
    "python-jose[cryptography]>=3.3.0",
    "pyyaml>=6.0.1",
    "python-dateutil>=2.8.2",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "black>=24.1.1",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/attilaszabo/iap-local-emulator"
Documentation = "https://github.com/attilaszabo/iap-local-emulator#readme"
Repository = "https://github.com/attilaszabo/iap-local-emulator.git"
Issues = "https://github.com/attilaszabo/iap-local-emulator/issues"

[project.scripts]
iap-emulator = "iap_emulator.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["iap_emulator*"]
exclude = ["tests*", "examples*", "docs*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=iap_emulator"
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]
python_classes = [
    "Test*",
]
python_functions = [
    "test_*",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests (require emulator running)",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["iap_emulator"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
precision = 2
show_missing = true

[tool.poetry]
package-mode = false